pretty large refactor to cleanup concepts
