--- apps/web/src/app/(profile)/[username]/_components/dashboard/_db_actions.ts
@@ -1,53 +1,0 @@
-'use server';
-
-import type {
-  AdventChallenges,
-  Difficulty,
-} from '~/app/(profile)/[username]/_components/dashboard/_domain_actions';
-import { prisma } from '@repo/db';
-
-export async function AdventRetrieveData(userId: string) {
-  const advent: AdventChallenges | null = await prisma.track.findFirst({
-    where: {
-      slug: 'advent-of-typescript-2023',
-    },
-    include: {
-      trackChallenges: {
-        orderBy: {
-          orderId: 'asc',
-        },
-        include: {
-          challenge: {
-            include: {
-              submission: {
-                where: {
-                  userId,
-                },
-              },
-            },
-          },
-        },
-      },
-      enrolledUsers: {
-        where: {
-          id: userId,
-        },
-        select: {
-          id: true,
-        },
-      },
-    },
-  });
-  return advent;
-}
-
-export async function DifficultyRetrieveData(userId: string) {
-  const data: Difficulty[] | null =
-    await prisma.$queryRaw`SELECT Difficulty, COUNT(Id) as TotalCompleted FROM (SELECT DISTINCT Difficulty, Challenge.Id FROM Submission JOIN Challenge ON Submission.challengeId = Challenge.Id WHERE Submission.userId = ${userId} AND IsSuccessful = 1) unique_query GROUP BY Difficulty `;
-  return data;
-}
-export async function SharedSolutionRetrieveData(userId: string) {
-  const data: Difficulty[] | null =
-    await prisma.$queryRaw`SELECT COUNT(SharedSolutionId) as TotalCompleted FROM (SELECT DISTINCT SharedSolutionId FROM SharedSolution JOIN Challenge ON SharedSolution.challengeId = Challenge.Id JOIN Vote ON SharedSolution.Id = Vote.SharedSolutionId WHERE SharedSolution.userId = ${userId} AND rootType = 'SHAREDSOLUTION') unique_query`;
-  return data;
-}
