+++ apps/web/src/app/(profile)/[username]/_components/dashboard/badges/_advent_badges.ts
@@ -1,0 +1,102 @@
+import type {AllBadges, BadgeFn} from "../_actions";
+import {prisma} from "@repo/db";
+
+export interface AdventChallenges {
+  trackChallenges: { challenge: { submission: { isSuccessful: boolean }[] } }[];
+}
+
+export type AotBadges =
+// eslint-disable-next-line @typescript-eslint/sort-type-constituents
+  'aot-2023-bronze' | 'aot-2023-silver' | 'aot-2023-gold' | 'aot-2023-platinum';
+
+export const adventBadgesFn: BadgeFn =
+  async ({
+           userId,
+           badges,
+         }: {
+    userId: string;
+    badges: AllBadges[];
+  }) => {
+    const advent: AdventChallenges | null = await AdventRetrieveData(userId);
+    await AdventChallengeFn(badges, advent);
+  };
+export async function AdventRetrieveData(userId: string) {
+  const advent: AdventChallenges | null = await prisma.track.findFirst({
+    where: {
+      slug: 'advent-of-typescript-2023',
+    },
+    include: {
+      trackChallenges: {
+        orderBy: {
+          orderId: 'asc',
+        },
+        include: {
+          challenge: {
+            include: {
+              submission: {
+                where: {
+                  userId,
+                },
+              },
+            },
+          },
+        },
+      },
+      enrolledUsers: {
+        where: {
+          id: userId,
+        },
+        select: {
+          id: true,
+        },
+      },
+    },
+  });
+  return advent;
+}
+export const AdventChallengeFn = async (badges: AllBadges[], advent: AdventChallenges | null) => {
+  // Advent Badge Logic
+  const numberOfAttemptedHolidayChallenges =
+    advent?.trackChallenges.filter((trackChallenge) => {
+      return (trackChallenge.challenge.submission?.length ?? 0) > 0;
+    }).length ?? 0;
+
+  if (numberOfAttemptedHolidayChallenges > 0) {
+    badges.push({
+      slug: 'aot-2023-bronze',
+      name: 'Advent of TypeScript 2023 Bronze',
+      shortName: 'Advent',
+    });
+  }
+
+  const numberOfCompletedHolidayChallenges =
+    advent?.trackChallenges.filter((trackChallenge) => {
+      return trackChallenge.challenge.submission?.some((submission) => submission.isSuccessful);
+    }).length ?? 0;
+
+  if (numberOfCompletedHolidayChallenges >= 5) {
+    badges.push({
+      slug: 'aot-2023-silver',
+      name: 'Advent of TypeScript 2023 Silver',
+      shortName: 'Advent',
+    });
+  }
+
+  if (numberOfCompletedHolidayChallenges >= 15) {
+    badges.push({
+      slug: 'aot-2023-gold',
+      name: 'Advent of TypeScript 2023 Gold',
+      shortName: 'Advent',
+    });
+  }
+
+  if (numberOfCompletedHolidayChallenges >= 25) {
+    badges.push({
+      slug: 'aot-2023-platinum',
+      name: 'Advent of TypeScript 2023 Platinum',
+      shortName: 'Advent',
+    });
+  }
+};
+
+
